% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdi_get_data.R
\docType{data}
\name{econ_data}
\alias{econ_data}
\title{Function to parse economic data}
\format{
An object of class \code{environment} of length 1.
}
\usage{
econ_data
}
\arguments{
\item{symbols}{Character vector of symbol names to download. Default: NULL, which will return all the symbols described in @details.}
}
\value{
Applicable time series data from FRED, stored as xts thanks to the
internal getSymbols.FRED function. Data to be stored in a new environment called econ_data.
}
\description{
The mdi_get_data() function currently retrieves data from FRED.

The 'default' series includes a range of indicators from Fundamental to
Behavioral and Catalyst categories. The sub-categories and their respective
tickers and descriptions are described in the Details section. Users can
specify their own character of symbols TODO - add
descriptions next to ticker
}
\details{
The sub-categories and their respective tickers and descriptions are
described below.\tabular{ll}{
   FUNDAMENTAL \tab  \cr
   Economic Trend: \tab  \cr
   OECDLOLITOAASTSAM \tab TODO: add description \cr
   ICSA \tab  \cr
   INDPRO \tab  \cr
   T10Y2Y \tab  \cr
   BAA10Y \tab  \cr
   --- \tab --- \cr
   Liquidity: \tab  \cr
   NFCI \tab  \cr
   --- \tab --- \cr
   Velocity: \tab  \cr
   M2V \tab  \cr
   TOTCI \tab  \cr
   M0263AUSM500NNBR \tab  \cr
   M0264AUSM500NNBR \tab  \cr
   BOGZ1FA895050005Q \tab  \cr
   --- \tab --- \cr
   --- \tab --- \cr
   BEHAVIORAL \tab  \cr
   Confirmation Bias: Surveys \tab  \cr
   UMCSENT \tab  \cr
   CSUSHPISA \tab  \cr
   SPCS20RSA \tab  \cr
   --- \tab --- \cr
   Representative Bias: Options Market \tab  \cr
   VIXCLS \tab  \cr
   VXVCLS \tab  \cr
   EVZCLS \tab  \cr
   THREEFYTP10 - \tab  \cr
   --- \tab --- \cr
   Cognitive Dissonance: \tab  \cr
   EMVOVERALLEMV \tab  \cr
   --- \tab --- \cr
   Loss Aversion: \tab  \cr
   --- \tab --- \cr
   Anchoring Bias: \tab  \cr
   --- \tab --- \cr
   Framing Bias: \tab  \cr
   --- \tab --- \cr
   Overconfidence: \tab  \cr
   --- \tab --- \cr
   --- \tab --- \cr
   CATALYST \tab  \cr
   Earnings Surprise: \tab  \cr
   --- \tab --- \cr
   Economic Surprise: \tab  \cr
   CFNAIMA3 \tab  \cr
   STLENI \tab  \cr
   STLFSI2 \tab  \cr
   --- \tab --- \cr
   Geopolitics: \tab  \cr
   WLEMUINDXD \tab  \cr
}
}
\examples{
\dontrun{
econ_data$mdi_get_data()
View(econ_data$NFCI)
head(econ_data$ICSA)
}

}
\keyword{datasets}
